requirements.txt
mediapipe==0.10.9
opencv-contrib-python-headless==4.8.1.78
numpy<2.0.0
streamlit>=1.28.0
pyserial>=3.5
protobuf>=3.20.0,<4.0.0
